requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pandas==1.5.3
numpy==1.23.5
requests==2.31.0
waitress==2.1.2
scikit-learn==1.2.2

# Lighter and Replit-friendly Surprise install
scikit-surprise @ git+https://github.com/NicolasHug/Surprise.git@master
